requirements.txt
# =============================================================================
# LIVKIT AGENTS FRAMEWORK
# =============================================================================
livekit-agents>=1.2.0
livekit-plugins-google>=1.0.0
livekit-plugins-noise-cancellation>=1.0.0
livekit-agents[tavus]~=1.0

# =============================================================================
# FACE RECOGNITION DEPENDENCIES
# =============================================================================
insightface>=0.7.3
opencv-python>=4.10.0.84
numpy>=1.26.4
Pillow>=10.4.0
onnxruntime>=1.16.0

# =============================================================================
# AI/ML DEPENDENCIES
# =============================================================================
tensorflow>=2.16.0
scikit-learn>=1.3.0

# =============================================================================
# DATA PROCESSING
# =============================================================================
pandas>=2.2.2
PyPDF2>=3.0.0

# =============================================================================
# COMMUNICATION & NOTIFICATIONS
# =============================================================================
requests>=2.31.0
smtplib-ssl>=1.0.0

# =============================================================================
# VOICE & AUDIO PROCESSING
# =============================================================================
speechrecognition>=3.10.0
pyaudio>=0.2.11

# =============================================================================
# WEB SEARCH & AI TOOLS
# =============================================================================
duckduckgo-search>=6.0.0
langchain-community>=0.3.0

# =============================================================================
# UTILITIES
# =============================================================================
python-dotenv>=1.0.0
tqdm>=4.66.4
gdown>=5.2.0
protobuf>=4.25.0

# =============================================================================
# OPTIONAL DEPENDENCIES (uncomment if needed)
# =============================================================================
# pyttsx3>=2.90  # Text-to-speech (not used in LiveKit version)
# mem0ai>=0.1.0  # Memory management (if needed)
